# Move InstKinds out to their own vector.