--- a/toolchain/sem_ir/inst.h
+++ b/toolchain/sem_ir/inst.h
@@ -196,8 +196,8 @@ class Inst : public Printable<Inst> {
     requires Internal::InstLikeType<TypedInst>
   auto As() const -> TypedInst {
     using Info = Internal::InstLikeTypeInfo<TypedInst>;
-    CARBON_CHECK(Is<TypedInst>(), "Casting inst {0} to wrong kind {1}", *this,
-                 Info::DebugName());
+    CARBON_DCHECK(Is<TypedInst>(), "Casting inst {0} to wrong kind {1}", *this,
+                  Info::DebugName());
     auto build_with_type_id_onwards = [&](auto... type_id_onwards) {
       if constexpr (Internal::HasKindMemberAsField<TypedInst>) {
         return TypedInst{kind(), type_id_onwards...};
@@ -454,6 +454,7 @@ class InstStore {
   // Reserves space.
   auto Reserve(size_t size) -> void {
     loc_ids_.reserve(size);
+    kinds_.reserve(size);
     types_and_args_.Reserve(size);
   }