@jonmeow 2024-12-13
Note, I'll lean on @chandlerc to decide whether this is worth it. The reduction in memory locality might be a net loss, performance versus space.