@zygoloid 2024-12-13
Yeah, this is what I was worried about -- a small performance decrease seems to be visible here, but we can probably get some performance benefits by avoiding looking at the operands of instructions in cases where we only need the kind. It's hard to know if that'll be a net benefit or regression unless we do the work to only look at the kind where possible.

Can you try changing `InstStore::Is` and `InstStore::TryGetAs` to look at the kind directly instead of first grabbing the whole `Inst`, and see if that makes any difference? (I'm not expecting a big change there, but you never know.)