@jonmeow 2024-12-16
@danakj To be sure, what we're looking at right now is the possibility of a significant enough performance regression that I would just close this PR.