@jonmeow 2024-12-16
Flipping a CHECK to DCHECK in `As` (on both trunk and branch for this comparison):

branch:

```
----------------------------------------------------------------------------------------------------------------------------
Benchmark                                                 Time             CPU   Iterations      Bytes      Lines     Tokens
----------------------------------------------------------------------------------------------------------------------------
BM_CompileAPIFileDenseDecls<Phase::Check>/256       1071900 ns      1071338 ns         1024 4.91628M/s 182.015k/s 1.02769M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/1024      1971624 ns      1969992 ns         1024 16.4919M/s 496.956k/s 2.91981M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/4096      5489374 ns      5484396 ns          256 25.8703M/s 734.447k/s 4.35034M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/16384    20271944 ns     20249973 ns           64 30.1704M/s 805.581k/s 4.77981M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/65536    82548931 ns     82498951 ns           16 30.5996M/s 793.465k/s 4.71047M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/262144  357791786 ns    357549384 ns            8 28.5831M/s 733.127k/s 4.35267M/s
```

trunk:

```
----------------------------------------------------------------------------------------------------------------------------
Benchmark                                                 Time             CPU   Iterations      Bytes      Lines     Tokens
----------------------------------------------------------------------------------------------------------------------------
BM_CompileAPIFileDenseDecls<Phase::Check>/256       1082382 ns      1081691 ns         1024 4.86923M/s 180.273k/s 1.01785M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/1024      1967895 ns      1966100 ns         1024 16.5246M/s  497.94k/s 2.92559M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/4096      5382641 ns      5379362 ns          256 26.3754M/s 748.788k/s 4.43528M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/16384    19875072 ns     19860513 ns           64  30.762M/s 821.379k/s 4.87354M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/65536    81334970 ns     81277123 ns           16 31.0596M/s 805.393k/s 4.78128M/s
BM_CompileAPIFileDenseDecls<Phase::Check>/262144  351240852 ns    350994095 ns            8 29.1169M/s 746.819k/s 4.43396M/s
```

This still looks like a 1-2% regression