@jonmeow 2024-12-16
I tried refactoring `GetAs` to avoid the full `Get` conversion, and it still seems not to affect performance. My leaning is now that this is going to remain a small hit to performance.

Looking at eval.cpp, I think it'd be hard to actually refactor things to avoid fetches in that case, and I suspect that'll be generally true. Closing, can reopen if someone has more ideas.